# perf(rust, python): csv-file use fast-float for csv float parsing