@aldanor 2023-03-11
@ritchie46 Do you know that fast-float crate has been since ported (with some tweaks, by none other than lexical's author himself) to rustc? :)

https://github.com/rust-lang/rust/tree/e350fe4e608b653da47e8012d13ef701613e717b/library/core/src/num/dec2flt

fast-float may end up being faster/slower compared to rustc's current implementation, but will be almost always faster than lexical, yea.