@ritchie46 2023-03-11
So benchmark again and simply use core? That's great news with regard to dependencies.