@aldanor 2023-03-11
@ritchie46 Yea, I'd definitely benchmark it against the core first, because it's largely just a port of fast-float now.

(there's actually quite a number of libraries out there that were using lexical pre-mid-2021 because it was so faster and are using it still, but now core is actually faster)