@ritchie46 2023-03-11
Ok, this is interesting. rust `str.parse` was definitely slower:

```
[]  ritchie46 /home/ritchie46/code/memcheck$ perf stat ./target/release/memcheck                                                                                                                                   (base) 
[src/main.rs:22] cont = 9291890

 Performance counter stats for './target/release/memcheck':

          5.429,86 msec task-clock                #    0,999 CPUs utilized          
               308      context-switches          #   56,723 /sec                   
                14      cpu-migrations            #    2,578 /sec                   
           205.323      page-faults               #   37,814 K/sec                  
    22.590.235.259      cycles                    #    4,160 GHz                    
    61.042.857.234      instructions              #    2,70  insn per cycle         
    10.983.541.883      branches                  #    2,023 G/sec                  
       103.580.277      branch-misses             #    0,94% of all branches        

       5,437846101 seconds time elapsed

       5,166718000 seconds user
       0,264343000 seconds sys


[]  ritchie46 /home/ritchie46/code/memcheck$ perf stat ./target/release/memcheck                                                                                                                                   (base) 
[src/main.rs:22] cont = 9291890

 Performance counter stats for './target/release/memcheck':

          5.849,62 msec task-clock                #    1,000 CPUs utilized          
               104      context-switches          #   17,779 /sec                   
                 9      cpu-migrations            #    1,539 /sec                   
           205.322      page-faults               #   35,100 K/sec                  
    22.473.453.954      cycles                    #    3,842 GHz                    
    61.036.475.833      instructions              #    2,72  insn per cycle         
    10.981.945.522      branches                  #    1,877 G/sec                  
       103.441.504      branch-misses             #    0,94% of all branches        

       5,850242729 seconds time elapsed

       5,610676000 seconds user
       0,240114000 seconds sys


[]  ritchie46 /home/ritchie46/code/memcheck$ perf stat ./target/release/memcheck                                                                                                                                   (base) 
[src/main.rs:22] cont = 9291890

 Performance counter stats for './target/release/memcheck':

          5.949,21 msec task-clock                #    1,000 CPUs utilized          
               112      context-switches          #   18,826 /sec                   
                10      cpu-migrations            #    1,681 /sec                   
           202.981      page-faults               #   34,119 K/sec                  
    22.393.802.084      cycles                    #    3,764 GHz                    
    61.026.180.157      instructions              #    2,73  insn per cycle         
    10.980.178.611      branches                  #    1,846 G/sec                  
       102.937.561      branch-misses             #    0,94% of all branches        

       5,949355466 seconds time elapsed

       5,678412000 seconds user
       0,271923000 seconds sys


[]  ritchie46 /home/ritchie46/code/memcheck$ perf stat ./target/release/memcheck                                                                                                                                   (base) 
[src/main.rs:22] cont = 9291890

 Performance counter stats for './target/release/memcheck':

          5.381,19 msec task-clock                #    1,000 CPUs utilized          
               144      context-switches          #   26,760 /sec                   
                16      cpu-migrations            #    2,973 /sec                   
           202.982      page-faults               #   37,721 K/sec                  
    22.389.937.992      cycles                    #    4,161 GHz                    
    61.026.043.630      instructions              #    2,73  insn per cycle         
    10.980.133.794      branches                  #    2,040 G/sec                  
       103.126.474      branch-misses             #    0,94% of all branches        

       5,381715314 seconds time elapsed

       5,125557000 seconds user
       0,256077000 seconds sys


```