@aldanor 2023-03-11
> Ok, this is interesting. rust str.parse was definitely slower

Because it has to deal with utf8 perhaps? 🤔 

Well... as I said, core parsers are based off fast-float but are not identical – I guess fast-float is the fastest option here then indeed :)