@ritchie46 2023-03-11
> Because it has to deal with utf8 perhaps?

Nah, I quickly benchmarked with `from_utf8_unchecked(bytes).parse()`.

> Well... as I said, core parsers are based off fast-float but are not identical – I guess fast-float is the fastest option here then indeed :)

:rocket: